x/citizen/types/message_set_approve_for_all_test.go
package types

import (
	"testing"

	"demo-onechain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSetApproveForAll_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetApproveForAll
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetApproveForAll{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetApproveForAll{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
